Remove the datadog metrics plugin
I have no means for testing this, and given that the kafka-connect-s3 plugin seems to be able to plug in
metrics anyways: Someone else feel free to resurrect this externally.

Right now this plugin will just simply start breaking without me noticing (probably), or in a way that
makes me spend time on just trying to keep things compiling.